optimized it
